requirements.txt
langchain
langchain-community
langsmith
openai
python-dotenv
streamlit
